martin-server/requirements.txt
aspy.yaml==1.0.0
astroid==1.5.3
bandit==1.4.0
billiard==3.5.0.3
boltons==16.1.1
brotlipy==0.7.0
cached-property==1.3.1
celery==4.1.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
coverage==4.4.2
decorator==4.1.2
dj-database-url==0.4.2
Django==2.0
django-log-request-id==1.3.2
django-model-utils==3.0.0
django-naomi==0.8
django-webpack-loader==0.5.0
dodgy==0.1.9
dparse==0.2.1
enum-compat==0.0.2
gunicorn==19.7.1
identify==1.0.7
idna==2.6
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
martin-lib==0.0.1
model-mommy==1.4.0
netifaces==0.10.6
nodeenv==1.2.0
opbeat==3.5.3
packaging==16.8
parso==0.1.0
pbr==3.1.1
pexpect==4.3.0
pickleshare==0.7.4
pre-commit==1.4.1
prompt-toolkit==1.0.15
prospector==0.12.7
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-decouple==3.1
pytz==2017.3
PyYAML==3.12
redis==2.10.6
requests==2.18.4
requirements-detector==0.5.2
safety==1.6.1
setoptconf==0.2.0
simplegeneric==0.8.1
six==1.11.0
smmap2==2.0.3
snowballstemmer==1.2.1
stevedore==1.27.1
termcolor==1.1.0
traitlets==4.3.2
urllib3==1.22
vine==1.1.4
vulture==0.26
wcwidth==0.1.7
whitenoise==3.3.1
wrapt==1.10.11
zeroconf==0.19.1
